backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-core
langchain-community
langchain-huggingface
langchain-text-splitters
huggingface-hub
sentence-transformers
transformers
torch
azure-search-documents
azure-identity
pydantic
python-dotenv
